+++ package.json
@@ -25,6 +25,7 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@types/uuid": "^9.0.1",
     "nodemon": "^2.0.19"
   },
   "nodemonConfig": {
