+++ server/utilities/generateUtils.ts
